+++ src/index.js
@@ -5,6 +5,7 @@ export const rules = {
   'namespace': require('./rules/namespace'),
   'no-namespace': require('./rules/no-namespace'),
   'export': require('./rules/export'),
+  'extensions': require('./rules/extensions'),
 
   'no-named-as-default': require('./rules/no-named-as-default'),
   'no-named-as-default-member': require('./rules/no-named-as-default-member'),